[Mobile] fixing diary
